messagesender/requirements.txt
pika==1.3.0
requests==2.22.0
pika-stubs==0.1.3
PyYAML==5.1.2